migrate Monitor class into MonitorCommand
